refactor(#792): split runtime_graph.rs into assembly/update/impl modules (sweep)
runtime_graph.rs 1279 -> 377 (leaves the debt list). The graph entry points
(build_runtime_graph / build_per_input_runtimes / build_chain_runtime_state) and
the RuntimeGraph struct stay; the rest moves by responsibility into siblings:

  - runtime_graph_assemble.rs  assemble_chain_runtime_state + the shared
                               per-segment helpers (build_input_processing_state,
                               build_output_routing_state, collect_bypass_block_ids,
                               output_entry_layout, target_for_route)
  - runtime_graph_update.rs    the in-place lock-free rebuild path
                               (update_chain_runtime_state[_spillover])
  - runtime_graph_impl.rs      RuntimeGraph methods (upsert/remove/runtimes_for)
  - runtime_graph_issue_736_tests.rs  the inline per-binding-rate test module

Pure mechanical move — no behavior change. The public fns are re-exported from
runtime_graph so engine::runtime_graph::* and the engine::runtime::* facade keep
resolving unchanged for every caller/test. group_segments_by_input and
input_group_ids become pub(crate) for the sibling modules.

Verified: engine suite green (volume/golden/isolation/deadline invariants pass),
adapter-gui + infra-cpal suites green, full workspace test targets compile clean,
zero warnings.
