+++ crates/engine/src/lib.rs
@@ -27,6 +27,9 @@ pub mod runtime_di_seek;
 pub mod runtime_dsp;
 pub mod runtime_endpoints;
 pub mod runtime_graph;
+mod runtime_graph_assemble;
+mod runtime_graph_impl;
+mod runtime_graph_update;
 pub mod runtime_io;
 pub mod runtime_layout;
 pub mod runtime_load;
