+++ scripts/validate.sh
@@ -67,7 +67,6 @@ crates/block-mod/src/lib_tests.rs
 crates/engine/src/audio_alloc_invariant_tests.rs
 crates/engine/src/audio_signal_integrity_tests.rs
 crates/engine/src/rig_runtime_tests.rs
-crates/engine/src/runtime_graph.rs
 crates/engine/src/runtime_tests.rs
 crates/engine/src/runtime.rs
 crates/engine/src/stream_isolation_tests.rs
